Re-order files a bit, validate continue URL, and remove old tests
