+++ src/login/controller/mfa.ts
@@ -1,37 +1,37 @@
 import Controller from '@curveball/controller';
 import { Context } from '@curveball/core';
 import querystring from 'querystring';
-
-import { mfaForm } from './formats/html';
-
-import log from '../log/service';
-import { EventType } from '../log/types';
-import * as userService from '../user/service';
-import { User } from '../user/types';
+import { isValidateRedirect } from '../utilities';
+import { mfaForm } from '../formats/html';
+import log from '../../log/service';
+import { EventType } from '../../log/types';
+import * as userService from '../../user/service';
+import { User } from '../../user/types';
 
 class MFAController extends Controller {
 
     async get(ctx: Context) {
-        let user: User = ctx.state.session.mfa_user;
 
-        if (!user) {
-            return this.redirectToLogin(ctx);
-        }
+      const user: User = ctx.state.session.mfa_user;
 
-        ctx.response.type = 'text/html';
-        ctx.response.body = mfaForm(
-          ctx.query.msg,
-          ctx.query.error,
-          {
-            continue: ctx.query.continue,
-          },
-        );
-
+      if (!user) {
+        return this.redirectToLogin(ctx);
       }
+
+      ctx.response.type = 'text/html';
+      ctx.response.body = mfaForm(
+        ctx.query.msg,
+        ctx.query.error,
+        {
+          continue: ctx.query.continue,
+        },
+      );
+
+    }
 
     async post(ctx: Context) {
 
-        let user: User = ctx.state.session.mfa_user;
+      const user: User = ctx.state.session.mfa_user;
 
         if (!user) {
             return this.redirectToLogin(ctx);
@@ -44,6 +44,10 @@ class MFAController extends Controller {
           }
         } else {
           return this.redirectToMfa(ctx, 'TOTP token required');
+        }
+
+        if (ctx.request.body.continue && !isValidateRedirect(ctx.request.body.continue)) {
+          return this.redirectToMfa(ctx, 'Invalid continue URL provided');
         }
 
         ctx.state.session = {
@@ -62,15 +66,15 @@ class MFAController extends Controller {
 
     async redirectToMfa(ctx: Context, error: string) {
 
-        ctx.response.status = 303;
-        ctx.response.headers.set('Location', '/mfa?' + querystring.stringify({ error }));
+      ctx.response.status = 303;
+      ctx.response.headers.set('Location', '/mfa?' + querystring.stringify({ error }));
 
     }
 
     async redirectToLogin(ctx: Context) {
 
-        ctx.response.status = 303;
-        ctx.response.headers.set('Location', '/login');
+      ctx.response.status = 303;
+      ctx.response.headers.set('Location', '/login');
 
     }
 }
