+++ src/login/utilities.ts
@@ -1,0 +1,4 @@
+export function isValidateRedirect(url: string): boolean {
+    // For now, /authorize is the only valid URL to redirect to
+    return url.startsWith('/authorize');
+}
