+++ src/routes.ts
@@ -7,8 +7,8 @@ import health from './health/controller';
 import home from './home/controller';
 import introspect from './introspect/controller';
 import userLog from './log/controller/user';
-import login from './login/controller';
-import mfa from './mfa/controller';
+import login from './login/controller/login';
+import mfa from './login/controller/mfa';
 import logout from './logout/controller';
 import oauth2Authorize from './oauth2/controller/authorize';
 import oauth2Token from './oauth2/controller/token';
